[ #3855 ] occursCheck: check for erased definitions
This created some issues with definitions created in the type world:
pattern lambdas (extended lambdas and absurd lambdas).
Since these definitions also enter terms, we cannot create them in @0.
For now, we always create them in @ω.

In the future, we might find ways to create a pattern lambda in @0 if it
is safe to do so or if the user explicitely tells us to.